swodlr-ui: small redux refactor
