+++ package.json
@@ -22,7 +22,6 @@
     "react-redux": "^8.0.5",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
-    "uuid": "^9.0.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
