+++ src/components/app/App.css
@@ -74,7 +74,6 @@ MAP
 
 .Map-container {
   height: 100%;
-  /* margin-top: 12px; */
   background-color: #c7c7c7;
   position: relative;
 }
@@ -93,7 +92,6 @@ NAVBAR
   height: 20px;
   width: 48px;
   margin-right: 10px;
-  /* cursor: 'pointer'; */
 }
 .dark-mode-switch .form-check-input:focus {
   border-color: rgba(0, 0, 0, 0.25);
@@ -103,7 +101,6 @@ NAVBAR
   cursor: 'pointer';
 }
 .dark-mode-switch .form-check-input:checked {
-  /* background-color: #2ab464ff; */
   background-color: #5C946E;
   border-color: #2ab464ff;
   border: none;
@@ -123,7 +120,6 @@ NAVBAR
 
 .dark-mode-container-background {
   background-color: #3D5D82;
-  /* background-color: #274c61; */
   color: #FAF9F6
 }
 
@@ -138,9 +134,6 @@ NAVBAR
 .light-mode-background {
   background-color:	#FAF9F6;
   color: black;
-  /* border-color: black;
-  border-style: solid;
-  border-width: thin; */
 }
 
 .light-mode-navbar-background {
@@ -163,9 +156,6 @@ NAVBAR
 
 .Main-navbar {  
   box-shadow: 0 2px 4px 0 rgba(0,0,0,.2);
-  /* background-color: #282c34; */
-  /* background-color: #1A2535; */
-  /* background-color: #02182B; */
 }
 
 /**************************************************
@@ -175,8 +165,6 @@ CUSTOMIZE PRODUCT CONTAINER
 .Customize-products-container-all {
   height: 93%;
   width: 75%;
-  /* background-color: #2276AD; */
-  /* background-color: #274c61ff; */
   margin: 20px;
   border-radius: 10px;
   position: absolute;
@@ -184,13 +172,10 @@ CUSTOMIZE PRODUCT CONTAINER
   left: 0;
   margin-top: 75px;
   z-index: 1000;
-  /* padding-top: 200px; */
-  /* overflow: hidden; */
 }
 
 .Customize-products-container {
   padding: 5px 25px 20px 25px;
-  /* overflow-y: auto; */
 }
 
 .Customize-products-container-min {
@@ -206,77 +191,38 @@ CUSTOMIZE PRODUCT CONTAINER
 }
 
 .Customize-products-container-sidebar-all {
-  /* height: 94.3%; */
   height: 100vh;
   width: 65%;
-  /* background-color: #2276AD; */
-  /* background-color: #274c61ff; */
-  /* margin: 15px; */
-  /* border-radius: 20px; */
   position: absolute;
   top: 0;
   left: 0;
-  /* margin-top: 55px;
-  margin-bottom: 55px; */
   z-index: 1000;
   padding-top: 60px;
-  /* overflow: hidden; */
-  /* background-color: #407899; */
-  /* background-color: #3D5D82; */
 }
 
 .sidebar-resize {
   margin: 0;
   position: absolute;
   background-color: #1a8754;
-  /* border-style: solid; */
-  /* border-color: #FAF9F6; */
   border-width: 1px;
   top: 50%;
   right: 0;
   margin-right: -14px;
-  /* border-radius: 30%; */
   height: 60px;
   width: 30px;
   cursor: col-resize;
 }
 
 .sidebar-resize-icon {
-  /* margin: 0; */
   position: relative;
   top: 15%;
-  /* right: 0; */
   height: 60%;
   cursor: col-resize;
-  /* margin-right: -11px; */
 }
 
 /**************************************************
 PRODUCT TABLE
 **************************************************/
-/* 
-.dark-mode-table .table-hover tbody tr:hover td, .table-hover tbody tr:hover th {
-  color: #FAF9F6;
-  cursor: pointer;
-}
-
-.light-mode-table .table-hover tbody tr:hover td, .table-hover tbody tr:hover th {
-  color: black;
-  cursor: pointer;
-  background-color: orange;
-}
-
-.dark-mode-table tr.hoverable-row tr:hover td:hover {
-  cursor: pointer;
-  background-color: orange;
-  color: white
-}
-
-.light-mode-table tr.hoverable-row tr:hover td:hover {
-  cursor: pointer;
-  background-color: orange;
-  color: white
-} */
 
 .dark-mode-table tr:hover td {
   /* background-color: pink; */
@@ -285,8 +231,6 @@ PRODUCT TABLE
 thead {
   position: sticky;
   top: -1px;
-  /* background-color: #2276AD; */
-  /* background-color: #274c61ff; */
 }
 
 .table-checkbox {
@@ -332,7 +276,6 @@ WELCOME PAGE
 **************************************************/
 
 .welcome-page-container {
-  /* background-color: #FAF9F6; */
   height: 100%;
   padding-top: 0px !important;
 }
@@ -347,20 +290,3 @@ WELCOME PAGE
   background-color: #5C946E;
   border-color: #5C946E;
 }
-
-/* .button-style:hover{
-  background-color: #5C946E;
-  border-color: #5C946E;
-  opacity: 50%;
-}
-
-.button-style:active{
-  background-color: #5C946E;
-  border-color: #5C946E;
-  opacity: 50%;
-}
-
-.button-style:disabled{
-  background-color: #5C946E;
-  border-color: #5C946E;
-} */
