+++ src/components/app/App.tsx
@@ -3,18 +3,18 @@ import MainNavbar from '../navbar/MainNavbar';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css';
 import Map from '../map/Map'
-import CustomizeProductsFooter from '../sidebar/CustomizeProductsFooter';
+import CustomizeProductsFooter from '../misc/CustomizeProductsFooter';
 import { useAppSelector, useAppDispatch } from '../../redux/hooks'
-import { setResizeInactive, setResizeEndLocation, setResizeStartLocation } from '../sidebar/customizeProductsFooterSlice';
-import CustomizeProductsFooterMin from '../sidebar/CustomizeProductsFooterMin';
+import { setResizeInactive, setResizeEndLocation, setResizeStartLocation } from '../sidebar/actions/sidebarSlice';
+import CustomizeProductsFooterMin from '../misc/CustomizeProductsFooterMin';
 import Welcome from '../welcome/Welcome'
 import CustomizeProductsSidebar from '../sidebar/CustomizeProductsSidebar';
 
 const App = () => {
   const dispatch = useAppDispatch()
-  const footerResizeActive = useAppSelector((state) => state.customizeProductsFooter.footerResizeActive)
-  const previousResizeEndLocation = useAppSelector((state) => state.customizeProductsFooter.resizeEndLocation)
-  const footerMinimized = useAppSelector((state) => state.customizeProductsFooter.footerMinimized)
+  const footerResizeActive = useAppSelector((state) => state.sidebar.footerResizeActive)
+  const previousResizeEndLocation = useAppSelector((state) => state.sidebar.resizeEndLocation)
+  const footerMinimized = useAppSelector((state) => state.sidebar.footerMinimized)
   // const userAuthenticated = useAppSelector((state) => state.app.userAuthenticated)
 
   const handleFooterResize = (event: any) => {
